requirements.txt
fastapi==0.45.0
mysql.connector
uvicone 